% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{flag_fillr_country}
\alias{flag_fillr_country}
\title{Use flags as a fill argument in a ggplot2 map}
\usage{
flag_fillr_country(country = "", resolution = c("small", "large"),
  size = c("100", "250", "1000"))
}
\arguments{
\item{country}{\code{character}. For a list of countries, run
\code{country_list()} in the R console.}
}
