% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_fillr.R, R/user_data.R
\name{flag_fillr_data}
\alias{flag_fillr_data}
\alias{flag_fillr_data}
\title{Use flags as a fill argument in a ggplot2 map}
\usage{
flag_fillr_data(data, partner_col = NULL, iso_codes = NULL,
  resolution = c("small", "large"), type = c("country", "state"),
  country = NULL, size = c("100", "250", "1000"), countries_col = NULL,
  state_col = NULL)

flag_fillr_data(data, partner_col = NULL, iso_codes = NULL,
  resolution = c("small", "large"), type = c("country", "state"),
  country = NULL, size = c("100", "250", "1000"), countries_col = NULL,
  state_col = NULL)
}
\description{
\code{flag_fillr_data} uses state-level flags as a fill for a
particular country.

\code{flag_fillr_data} uses state-level flags as a fill for a
particular country.
}
\examples{
\dontrun{
flag_fillr_data()
}
\dontrun{
flag_fillr_data()
}
}
