% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{flag_fillr_country}
\alias{flag_fillr_country}
\title{Use flags to plot a ggplot2 map}
\usage{
flag_fillr_country(country = NULL, resolution = c("small", "large"),
  size = c("100", "250", "1000"))
}
\arguments{
\item{country}{\code{character}. For a list of countries, run
\code{country_list()} in the R console.}

\item{resolution}{detail of rnaturalearth data.}

\item{size}{size of flag image.}
}
\description{
Plot a map of a country using flags as fills.
}
