% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_fillr.R
\name{flag_fillr_data}
\alias{flag_fillr_data}
\title{Use flags as a fill argument in a ggplot2 map}
\usage{
flag_fillr_data(data, partner_col = NULL, resolution = c("small", "large"),
  type = c("countries", "states"), country = NULL, countries_col = NULL,
  states_col = NULL)
}
\description{
\code{flag_fillr_data} uses state-level flags as a fill for a
particular country.
}
\examples{
\dontrun{
flag_fillr_data()
}
}
