% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data.R
\name{flag_fillr_data}
\alias{flag_fillr_data}
\title{Use flags as a fill argument in a ggplot2 map}
\usage{
flag_fillr_data(data = NULL, country = NULL, partner_col = NULL,
  resolution = c("small", "large"), type = c("country", "state"),
  size = c("100", "250", "1000"), state_col = NULL, country_col = NULL,
  mainland_only = TRUE)
}
\arguments{
\item{data}{A \code{data.frame}, see details.}

\item{country}{country name. If \code{country} is \code{NULL}, you must supply a \code{country_col}.}

\item{partner_col}{column of partner countries. See details.}

\item{resolution}{detail of rnaturalearth data.}

\item{type}{will you be plotting flags on countries or inside countries, on the states?}

\item{size}{size of flag image.}

\item{state_col}{column of state names. Must be full names, see details.}

\item{country_col}{column of country names. See details.}

\item{mainland_only}{remove non-contiguous landmasses? Will remove places
like Alaska, Hawaii, etc.}
}
\description{
\code{flag_fillr_data} uses user-supplied data to get flags as a fill for a map of a
particular country.
}
\details{
This function accepts a data.frame that must contain at least two columns:
one for the base territory (i.e. the area that will be mapped), and another
for the names of the countries whose flags will be used as fills in the map.
If the goal is to fill states inside of one country, the base territory column
should be called "state"; if you want to map entire countries with one fill,
it should be called "country". The column of the flags to be used should be called
"partner".
If you would rather not rename the columns in your data, you can specify the
'partner' (\code{partner_col}), 'state' (\code{state_col}), and 'country'
(\code{country_col}) columns. These must be the actual columns, written in
the format \code{df$state}, where \code{df} is your data.frame and
\code{state} is the 'state' column (with whatever name you have for it). For
information on acceptable country names, see \code{country_list()}.
}
\examples{
\dontrun{
 us_data <- data.frame(
   state = state.name, # I love R!
   partner = c(rep("Mexico", 15), rep("Canada", 25), rep("China", 8), rep("Japan", 2)),
   stringsAsFactors = FALSE
  )
  flag_fillr_data(us_data, country = "United States of America", type = "state", size = "250")
}
}
