% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data.R
\name{flag_fillr_data}
\alias{flag_fillr_data}
\title{Use flags as a fill argument in a ggplot2 map}
\usage{
flag_fillr_data(data = NULL, country = NULL, partner_col = NULL,
  resolution = c("small", "large"), type = c("country", "state"),
  size = c("100", "250", "1000"), state_col = NULL, country_col = NULL)
}
\description{
\code{flag_fillr_data} uses user-supplied data to get flags as a fill for a map of a
particular country.
}
\details{
This function accepts a data.frame that must contain at least two columns: one for
the base territory (i.e. the area that will be mapped), and another for the names of the countries
whose flags will be used as fills in the map. If the goal is to fill states inside of one country,
the base territory column should be called "state"; if you want to map entire countries with one fill,
it should be called "country". The column of the flags to be used should be called "partner".
If you would rather not rename the columns in your data, you can specify the 'partner'
(\code{partner_col}), 'state' (\code{state_col}), and 'country' (\code{country_col}) columns.
These must be the actual columns, written in the format \code{df$state}, where \code{df} is your
data.frame and \code{state} is the 'state' column (with whatever name you have for it).
}
\examples{
\dontrun{
df <- data.frame(
 state = c("Drenthe", "Flevoland", "Frisian", "Gelderland", "Groningen",
           "Limburg", "Noord-Brabant", "Noord-Holland", "Overijissel",
           "Utrecht", "Zeeland", "Zuid-Holland"),
 partner = c("Germany", "United Kingdom", "Germany", "United States",
             "Denmark", "Norway", "Sweden", "Germany", "Belgium",
             "France", "United Kingdom", "France"),
 stringsAsFactors = FALSE
)
flag_fillr_data(df, country = "Netherlands", size = "250")
}
}
